refactor: change dir
