+++ .vscode/launch.json
@@ -9,7 +9,7 @@
       "type": "go",
       "request": "launch",
       "mode": "auto",
-      "program": "${workspaceRoot}/",
+      "program": "${workspaceRoot}/src",
       "args": ["version"],
       "output": "${workspaceRoot}/build/garage",
       "trace": "verbose",
@@ -22,78 +22,8 @@
       "mode": "auto",
       "showLog": true,
       "debugAdapter": "dlv-dap",
-      "program": "${workspaceRoot}/",
+      "program": "${workspaceRoot}/src",
       "args": ["help"],
-      "output": "${workspaceRoot}/build/garage",
-      "trace": "verbose",
-      "cwd": "${workspaceRoot}"
-    },
-    {
-      "name": "Command Code Help",
-      "type": "go",
-      "request": "launch",
-      "mode": "auto",
-      "showLog": true,
-      "debugAdapter": "dlv-dap",
-      "program": "${workspaceRoot}/",
-      "args": ["code", "--help"],
-      "output": "${workspaceRoot}/build/garage",
-      "trace": "verbose",
-      "cwd": "${workspaceRoot}"
-    },
-    {
-      "name": "Command Code",
-      "type": "go",
-      "request": "launch",
-      "mode": "auto",
-      "showLog": true,
-      "debugAdapter": "dlv-dap",
-      "program": "${workspaceRoot}/",
-      "args": ["jav_code", "--proxy=http://127.0.0.1:7890", "PCDE-016"],
-      "output": "${workspaceRoot}/build/garage",
-      "trace": "verbose",
-      "cwd": "${workspaceRoot}"
-    },
-    {
-      "name": "Command Subtitle Help",
-      "type": "go",
-      "request": "launch",
-      "mode": "auto",
-      "showLog": true,
-      "debugAdapter": "dlv-dap",
-      "program": "${workspaceRoot}/",
-      "args": ["video_import_subtitle", "--help"],
-      "output": "${workspaceRoot}/build/garage",
-      "trace": "verbose",
-      "cwd": "${workspaceRoot}"
-    },
-    {
-      "name": "Command Subtitle",
-      "type": "go",
-      "request": "launch",
-      "mode": "auto",
-      "showLog": true,
-      "debugAdapter": "dlv-dap",
-      "program": "${workspaceRoot}/",
-      "args": [
-        "video_import_subtitle",
-        "--source_root_path=testdata",
-        "--dest_path=testdata/result",
-        "--fonts_path=testdata/fonts"
-      ],
-      "output": "${workspaceRoot}/build/garage",
-      "trace": "verbose",
-      "cwd": "${workspaceRoot}"
-    },
-    {
-      "name": "Command Convert",
-      "type": "go",
-      "request": "launch",
-      "mode": "auto",
-      "showLog": true,
-      "debugAdapter": "dlv-dap",
-      "program": "${workspaceRoot}/",
-      "args": ["video_convert", "--source_root_path=testdata"],
       "output": "${workspaceRoot}/build/garage",
       "trace": "verbose",
       "cwd": "${workspaceRoot}"
