+++ Makefile
@@ -18,20 +18,20 @@ all: release_linux release_win release_mac
 release_linux:
 	# Build for linux
 	go clean
-	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -v -trimpath  -ldflags ${ldflags} -o ${BuildDIR}/${APP}-linux64-amd64 .
-	CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-linux64-arm64 .
+	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -v -trimpath  -ldflags ${ldflags} -o ${BuildDIR}/${APP}-linux64-amd64 ./src
+	CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-linux64-arm64 ./src
 
 release_win:
 	# Build for win
 	go clean
-	CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-windows-amd64.exe .
-	CGO_ENABLED=0 GOOS=windows GOARCH=arm go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-windows-arm.exe .
+	CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-windows-amd64.exe ./src
+	CGO_ENABLED=0 GOOS=windows GOARCH=arm go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-windows-arm.exe ./src
 
 release_mac:
 	# Build for mac
 	go clean
-	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-darwin-amd64 .
-	CGO_ENABLED=0 GOOS=darwin GOARCH=arm64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-darwin-arm64 .
+	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-darwin-amd64 ./src
+	CGO_ENABLED=0 GOOS=darwin GOARCH=arm64 go build -v -trimpath -ldflags ${ldflags} -o ${BuildDIR}/${APP}-darwin-arm64 ./src
 
 clean:
 	@go clean --cache
