+++ src/cmd/jav_cmd/star.go
