+++ src/cmd/version.go
